requirements.txt
numpy==1.24.3
packaging==23.1
pandas==2.0.2
patsy==0.5.3
plotly==5.15.0
plotly-express==0.4.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
scipy==1.10.1
six==1.16.0
statsmodels==0.14.0
tenacity==8.2.2
tzdata==2023.3
